Refactored chemical reaction network
It is now in 3 files:
species, reaction, propensities, and chemical_reaction_network

Interim commit with regards to setters etc for Propensities
