+++ biocrnpyler/mechanisms_enzyme.py
@@ -1,6 +1,7 @@
 from warnings import warn
 from .mechanism import *
-from .chemical_reaction_network import Species, Reaction, ComplexSpecies, Multimer
+from .species import Species, ComplexSpecies, Multimer
+from .reaction import Reaction
 
 
 class BasicCatalysis(Mechanism):
