+++ biocrnpyler/mixtures_extract.py
@@ -11,7 +11,8 @@ from .mechanisms_binding import One_Step_Binding
 from .mechanisms_txtl import Transcription_MM, Translation_MM, Degredation_mRNA_MM, OneStepGeneExpression, SimpleTranscription, SimpleTranslation
 from .global_mechanism import Dilution
 from .mixture import Mixture
-from .chemical_reaction_network import Species, ChemicalReactionNetwork
+from .chemical_reaction_network import ChemicalReactionNetwork
+from .species import Species
 from .dna_assembly import DNAassembly
