+++ biocrnpyler/propensities.py
@@ -3,8 +3,9 @@
 
 from typing import Set, Union, Dict, List
 from numbers import Real
-from .parameter import Parameter
-from .sbmlutil import getSpeciesByName
+from .parameter import Parameter, ModelParameter, ParameterEntry
+from .sbmlutil import getSpeciesByName, _create_ratelaw_parameter
+from .species import Species
 import libsbml
 
 
@@ -56,12 +57,62 @@ class Propensity(object):
             rate_coeff_name = 'k'
         return rate_coeff_name
 
-    def pretty_print(self, **kwargs):
-        raise NotImplementedError
+    def _check_parameter(self, parameter, allow_None = False, positive = True):
+        """
+        A helper function used in setters to set parameters and do type checking
+        """
+        if isinstance(parameter, Parameter) and (parameter.value > 0 or not positive):
+            return parameter
+        elif (isinstance(parameter, float) or isinstance(parameter, int)) and (parameter > 0 or not positive):
+            return parameter
+        elif parameter is None and allow_None:
+            return parameter
+        else:
+            if positive:
+                raise ValueError(f"Propensity parameters must be Parameters or floats with positive values. Recieved {parameter}.")
+            else:
+                raise ValueError(f"Propensity parameters must be Parameters or floats. Recieved {parameter}.")
+
+    def _check_species(self, species, allow_None = False):
+        """
+        A helper function used in setters to set species and do type checking
+        """
+        if isinstance(species, Species):
+            return species
+        elif species is None and allow_None:
+            return species
+        else:
+            raise ValueError(f"Propensity expected a Species: recieved {species}.")
+
+
+    def pretty_print(self, show_parameters = True, **kwargs):
+        txt += self.pretty_print_rate(**kwargs)
+        if show_parameters:
+            txt += "\n"+self.pretty_print(parameters)
+        return txt
+
+    def pretty_print_rate(self, **kwargs):
+        raise NotImplementedError("class Propensity is meant to be subclassed!")
+
+    def pretty_print_parameters(self, show_keys = True, **kwargs):
+        txt = ""
+        for k in self.propensity_dict["parameters"]:
+            p = self.propensity_dict["parameters"][k]
+            if isinstance(p, Parameter):
+                txt += f"\t{k}={p.value}"#p.pretty_print(**kwargs)+"\n"
+                if isinstance(p, ModelParameter) and show_keys:
+                    txt+=f"\n\t\tfound_key={p.found_key}. search_key={p.search_key}."
+                txt+="\n"
+            else:
+                txt += f"\t{k}={p}\n"
+        return txt
 
     @property
     def is_reversible(self):
-        raise NotImplementedError
+        """
+        By default, Propensities are assumed to NOT be reversible.
+        """
+        return False
 
     @property
     def k_forward(self):
@@ -81,7 +132,7 @@ class Propensity(object):
         return self.propensity_dict['species'].values()
 
     def create_kinetic_law(self, reaction, reverse_reaction, stochastic):
-        raise NotImplementedError
+        raise NotImplementedError("class Propensity is meant to be subclassed!")
 
     @classmethod
     def from_dict(cls, propensity_dict):
@@ -111,29 +162,27 @@ class MassAction(Propensity):
 
     @property
     def k_forward(self):
-        return self._k_forward
+        if isinstance(self._k_forward, Parameter):
+            return self._k_forward.value
+        else:
+            return self._k_forward
 
     @k_forward.setter
     def k_forward(self, new_k_forward):
-        if new_k_forward <= 0:
-            raise ValueError(f'Forward reaction rate coefficient is negative! '
-                             f'k={new_k_forward}')
-
-        self._k_forward = new_k_forward
+        self._k_forward = self._check_parameter(new_k_forward)
+        self.propensity_dict['parameters']['k_forward'] = self.k_forward
 
     @property
     def k_reverse(self):
-        return self._k_reverse
+        if isinstance(self._k_reverse, Parameter):
+            return self._k_reverse.value
+        else:
+            return self._k_reverse
 
     @k_reverse.setter
     def k_reverse(self, new_k_reverse):
-        if not isinstance(self, MassAction):
-            # only mass action object can have reverse rates
-            raise NotImplementedError
-        if new_k_reverse is not None and new_k_reverse <= 0:
-            raise ValueError(f'Reverse reaction rate coefficient is negative! '
-                             f'k={new_k_reverse}')
-        self._k_reverse = new_k_reverse
+        self._k_reverse = self._check_parameter(new_k_reverse, allow_None = True)
+        self.propensity_dict['parameters']['k_reverse'] = self.k_reverse
 
     @property
     def is_reversible(self):
@@ -142,7 +191,7 @@ class MassAction(Propensity):
         else:
             return True
 
-    def pretty_print(self, **kwargs):
+    def pretty_print_rate(self, **kwargs):
         if 'reaction_direction' in kwargs and kwargs['react_direction'] is 'reverse':
             if self.is_reversible:
                 return f''
@@ -169,10 +218,8 @@ class MassAction(Propensity):
 
         # get the name of the kinetic rate coefficient
         rate_coeff_name = Propensity._get_parameter_name(self.k_forward)
-        param = ratelaw.createParameter()
-        param.setId(rate_coeff_name)
-        param.setConstant(True)
-        param.setValue(self.k_forward)
+        param = _create_ratelaw_parameter(ratelaw, rate_coeff_name, self.k_forward)
+
         # annotation_dict["k"] = rate_coeff
 
         rate_formula = self._get_rate_formula(rate_coeff_name, stochastic, reactant_species)
@@ -187,11 +234,8 @@ class MassAction(Propensity):
                 reactant_species[species_id] = w_species
 
             rate_coeff_name = Propensity._get_parameter_name(self.k_reverse)
-            param = ratelaw.createParameter()
-            param.setId(rate_coeff_name)
-            param.setConstant(True)
-            param.setValue(self.k_reverse)
-            # annotation_dict["k"] = rate_coeff
+            param = _create_ratelaw_parameter(ratelaw, rate_coeff_name, self.k_reverse)
+
             rate_formula += self._get_rate_formula(rate_coeff_name, stochastic, reactant_species)
         else:
             sbml_reaction.setReversible(False)
@@ -223,20 +267,72 @@ class MassAction(Propensity):
         return ratestring
 
 
-class Hill(MassAction):
-    def __init__(self, k_forward: float, s1, K: float, n: float):
-        super(Hill, self).__init__(k_forward=k_forward)
-        
+class Hill(Propensity):
+    def __init__(self, k: float, s1: Species, K: float, n: float, d:float):
+        Propensity.__init__(self)
         self.s1 = s1
         self.K = K
         self.n = n
-        self.propensity_dict['species']['s1'] = self.s1
+        self.d = d
+
+    @property
+    def k(self):
+        if isinstance(self._k, Parameter):
+            return self._k.value
+        else:
+            return self._k
+    @k.setter
+    def k(self, new_k):
+        self._k = self._check_parameter(new_k)
+        self.propensity_dict['parameters']['k'] = self.k
+
+    @property
+    def K(self):
+        if isinstance(self._K, Parameter):
+            return self._K.value
+        else:
+            return self._K
+
+    @K.setter
+    def K(self, new_K):
+        self._K = self._check_parameter(new_K)
         self.propensity_dict['parameters']['K'] = self.K
+
+    @property
+    def n(self):
+        if isinstance(self._n, Parameter):
+            return self._n.value
+        else:
+            return self._n
+    @n.setter
+    def n(self, new_n):
+        self._n = self._check_parameter(new_n)
         self.propensity_dict['parameters']['n'] = self.n
 
+    @property
+    def s1(self):
+        return self._s1
+    @s1.setter
+    def s1(self, new_s1):
+        self._s1 = self._check_species(new_s1)
+        self.propensity_dict['species']['s1'] = self.s1
+
+    @property
+    def d(self):
+        return self._d
+    @d.setter
+    def d(self, new_d):
+        self._d = self._check_species(new_d, allow_None = True)
+        self.propensity_dict['species']['d'] = self.d
+
+    def pretty_print_rate(self, show_parameters = True, **kwargs):
+        raise NotImplementedError("Propensity class Hill is meant to be subclassed: try HillPositive, HillNegative, ProportionalHillPositive, or ProportionalHillNegative.")
+
     def create_kinetic_law(self, model, sbml_reaction, stochastic, **kwargs):
+        raise NotImplementedError("Propensity class Hill is meant to be subclassed: try HillPositive, HillNegative, ProportionalHillPositive, or ProportionalHillNegative.")
 
-
+    
+    def create_kinetic_law(self, model, sbml_reaction, stochastic, **kwargs):
         param_n = ratelaw.createParameter()
         param_n.setId("n")
         param_n.setConstant(True)
@@ -247,6 +343,15 @@ class Hill(MassAction):
         param_K.setConstant(True)
         param_K.setValue(self.K)
 
+        param_k = ratelaw.createParameter()
+        parak_k.setId("k")
+        param_K.setConstant(True)
+        pram_K.setValue(self.k)
+
+        s = str(self.s1).replace("'", "")
+        s_species_id = getSpeciesByName(model,s).getId()
+
+
         # annotation_dict["k"] = self.k_forward
         # annotation_dict["K"] = self.K
         # annotation_dict["n"] = self.n
@@ -254,9 +359,6 @@ class Hill(MassAction):
         #
         #     #Create ratestring for non-massaction propensities
         # if propensity_type == "hillpositive":
-        #
-        #     s = str(propensity_params['s1']).replace("'", "")
-        #     s_species_id = getSpeciesByName(model,s).getId()
         #
         #     ratestring+=f"*{s_species_id}^n/({s_species_id}^n+K)"
         #
@@ -295,8 +397,9 @@ class Hill(MassAction):
         #     annotation_dict["d"] = d_species_id
 
 
+
 class HillPositive(Hill):
-    def __init__(self, k_forward: float, s1, K: float, n: float):
+    def __init__(self, k: float, s1:Species, K: float, n: float):
         """ Hill positive propensity is a nonlinear propensity with the following formula
 
             p(s1; k, K, n) = k*s1^n/(s1^n + K)
@@ -305,17 +408,15 @@ class HillPositive(Hill):
         :param s1: species (chemical_reaction_network.species)
         :param K: dissociation constant (float)
         """
-        super(HillPositive, self).__init__(k_forward=k_forward, s1=s1, K=K, n=n)
+        Hill.__init__(self = self, k=k, s1=s1, K=K, n=n, d=None)
 
-    def pretty_print(self,**kwargs):
-        if 'reaction_direction' in kwargs and kwargs['reaction_direction'] is 'reverse':
-            return super().pretty_print(**kwargs)
-
-        return f'k_f({self.s1}) = {self.k_forward}*{self.s1}^{self.n}/({self.s1}^{self.n} + {self.K})'
+    def pretty_print_rate(self, show_parameters = True, **kwargs):
+        return f'k_f({self.s1.pretty_print(**kwargs)}) = k {self.s1.pretty_print(**kwargs)}^n / ({self.s1.pretty_print(**kwargs)}^n + K)'
+        
 
 
 class HillNegative(Hill):
-    def __init__(self, k_forward: float, s1, K: float, n: float):
+    def __init__(self, k: float, s1, K: float, n: float):
         """ Hill negative propensity is a nonlinear propensity with the following formula
 
             p(s1; k, K, n) = k*1/(s1^n + K)
@@ -325,14 +426,14 @@ class HillNegative(Hill):
         :param K: dissociation constant (float)
         :param n: cooperativity (float)
         """
-        super(HillNegative, self).__init__(k_forward=k_forward, s1=s1, K=K, n=n)
+        Hill.__init__(self = self, k=k, s1=s1, K=K, n=n, d=None)
 
-    def pretty_print(self,**kwargs):
-        return f'k_f({self.s1}) = {self.k_forward}/({self.s1}^{self.n} + {self.K})'
+    def pretty_print_rate(self, show_parameters = True, **kwargs):
+        return f'k_f({self.s1.pretty_print(**kwargs)}) = k / ({self.s1.pretty_print(**kwargs)}^n + K)'
 
 
 class ProportionalHillPositive(HillPositive):
-    def __init__(self, k_forward: float, s1, K: float, n: float, d):
+    def __init__(self, k: float, s1:Species, K: float, n: float, d:Species):
         """ proportional Hill positive propensity with the following formula
 
             p(s1, d; k, K, n) = k*d*s1^n/(s1^n + K)
@@ -343,16 +444,14 @@ class ProportionalHillPositive(HillPositive):
         :param n: cooperativity (float)
         :param d: species (chemical_reaction_network.species)
         """
-        super(ProportionalHillPositive, self).__init__(k_forward=k_forward, s1=s1, K=K, n=n)
-        self.d = d
-        self.propensity_dict['species']['d'] = self.d
+        Hill.__init__(self = self, k=k, s1=s1, K=K, n=n, d=d)
 
-    def pretty_print(self, **kwargs):
-        return '*'.join([str(self.d), super(ProportionalHillPositive, self).pretty_print(**kwargs)])
+    def pretty_print_rate(self, show_parameters = True,  **kwargs):
+        return f'k_f({self.s1.pretty_print(**kwargs)}, {self.d.pretty_print(**kwargs)}) = k {self.d.pretty_print(**kwarks)} {self.s1.pretty_print(**kwargs)}^n/({self.s1.pretty_print(**kwargs)}^n + K)'
 
 
 class ProportionalHillNegative(HillNegative):
-    def __init__(self, k_forward: float, s1, K: float, n: float, d):
+    def __init__(self, k_forward: float, s1:Species, K: float, n: float, d:Species):
         """ proportional Hill negative propensity with the following formula
 
             p(s1, d; k, K, n) = k*d/(s1^n + K)
@@ -363,9 +462,7 @@ class ProportionalHillNegative(HillNegative):
         :param n: cooperativity (float)
         :param d: species (chemical_reaction_network.species)
         """
-        super(ProportionalHillNegative, self).__init__(k_forward=k_forward, s1=s1, K=K, n=n)
-        self.d = d
-        self.propensity_dict['species']['d'] = self.d
+        Hill.__init__(self = self, k=k, s1=s1, K=K, n=n, d=d)
 
-    def pretty_print(self,**kwargs):
-        return '*'.join([str(self.d), super(ProportionalHillNegative, self).pretty_print(**kwargs)])
+    def pretty_print(self, show_parameters = True, **kwargs):
+        return f'k_f({self.s1.pretty_print(**kwargs)}, {self.d.pretty_print(**kwargs)}) = k {self.d.pretty_print(**kwargs)} /({self.s1.pretty_print(**kwargs)}^{self.n} + K)'
