+++ pytest.ini
@@ -4,4 +4,4 @@ minversion = 5.4
 addopts = --nbval-lax --cov=biocrnpyler
 testpaths =
     Tests
-    examples
+#    examples
